feat: platform piece settings
