+++ packages/react-ui/src/features/pieces/lib/pieces-api.ts
@@ -8,15 +8,19 @@ import {
 import {
   Action,
   ActionType,
+  AddPieceRequestBody,
   GetPieceRequestParams,
   GetPieceRequestQuery,
   ListPiecesRequestQuery,
+  PackageType,
   PieceOptionRequest,
+  PieceScope,
   Trigger,
   TriggerType,
 } from '@activepieces/shared';
 
 import { StepMetadata } from './pieces-hook';
+import { authenticationSession } from '@/lib/authentication-session';
 
 export const PRIMITIVE_STEP_METADATA = {
   [ActionType.CODE]: {
@@ -97,7 +101,15 @@ export const piecesApi = {
       }
     }
   },
-  installCommunityPiece(params: FormData) {
+  install(params: AddPieceRequestBody) {
+    const formData = new FormData();
+    formData.set('packageType', params.packageType);
+    formData.set('pieceName', params.pieceName);
+    formData.set('pieceVersion', params.pieceVersion);
+    formData.set('scope', PieceScope.PROJECT);
+    if (params.packageType === PackageType.ARCHIVE) {
+      formData.set('pieceArchive', params.pieceArchive as any);
+    }
     return api.post<PieceMetadataModel>(`/v1/pieces`, params);
   },
   delete(id: string) {
